Separate out files for public API functions
