+++ python/rascaline/rascaline/utils/clebsch_gordan/__init__.py
@@ -1,7 +1,10 @@
-from .clebsch_gordan import correlate_density, correlate_density_metadata  # noqa
+from .correlate_density import correlate_density, correlate_density_metadata  # noqa
+from .correlate_tensors import correlate_tensors, correlate_tensors_metadata  # noqa
 
 
 __all__ = [
     "correlate_density",
     "correlate_density_metadata",
+    "correlate_tensors",
+    "correlate_tensors_metadata",
 ]
